WICKET-6786: Create OriginBasedResourceIsolationPolicy to support legacy browsers that don't send Sec-Fetch-* headers and add it as a default Resource Isolation Policy to the Fetch Metadata listener.
